requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.36
alembic==1.13.3
aiosqlite==0.19.0

# Caching
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
slowapi==0.1.9
bcrypt==4.0.1
cryptography==41.0.7

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.0.3

# External APIs
openai==1.3.5
google-generativeai==0.3.2
googlemaps==4.10.0
twilio==8.10.0
sendgrid==6.10.0
httpx==0.25.2
geopy==2.4.0
phonenumbers==8.13.27

# Document Processing
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.1.0
reportlab==4.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
hypothesis==6.88.1

# Error Tracking & Monitoring
sentry-sdk==1.39.1

# Cloud Services (Optional)
boto3==1.34.0  # AWS SDK for S3 backups and Secrets Manager
azure-keyvault-secrets==4.7.0  # Azure Key Vault
azure-identity==1.15.0  # Azure authentication

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1